Arrays, primitives, strings restructure
