+++ Arrays, Primitives, Strings/LongestPalindromeConstruction/WithHashtable.java
@@ -1,4 +1,4 @@
-class WithHashtable {
+class Solution {
   public int longestPalindrome(String s) {
     int longestPalindromeLength = 0;
 
@@ -48,29 +48,6 @@ class WithHashtable {
           longestPalindromeLength += count - 1;
         }
       }
-    }
-
-    return longestPalindromeLength;
-  }
-}
-
-class WithSet {
-  public int longestPalindrome(String s) {
-    int matchings = 0;
-    Set<Character> unmatchedCharacters = new HashSet<Character>();
-
-    for (char c: s.toCharArray()) {
-      if (unmatchedCharacters.contains(c)) {
-        unmatchedCharacters.remove(c);
-        matchings++;
-      } else {
-        unmatchedCharacters.add(c);
-      }
-    }
-
-    int longestPalindromeLength = matchings * 2;
-    if (!unmatchedCharacters.isEmpty()) {
-      longestPalindromeLength++; // place 1 unmatched odd character in the middle
     }
 
     return longestPalindromeLength;
