+++ Arrays, Primitives, Strings/RotateImage/LayerByLayer.java
@@ -1,4 +1,4 @@
-class LayerByLayerWithTemps {
+class Solution {
   public void rotate(int[][] matrix) {
     int size = matrix.length - 1; // this is really the last index
 
@@ -14,36 +14,6 @@ class LayerByLayerWithTemps {
         matrix[i][size - layer] = topFence;               // set value walking right fence
         matrix[size - layer][size - i] = rightFence;      // set value walking bottom fence
         matrix[size - i][layer] = bottomFence;            // set value walking left fence
-      }
-    }
-  }
-}
-
-class MatrixTranspose {
-  public void rotate(int[][] matrix) {
-    flipVertically(matrix);
-    transpose(matrix);
-  }
-
-  private void flipVertically(int[][] matrix) {
-    for (int topRow = 0; topRow < (matrix.length / 2); topRow++) {
-      int bottomRow = matrix.length - 1 - topRow;
-
-      int[] temp = matrix[topRow];
-      matrix[topRow] = matrix[bottomRow];
-      matrix[bottomRow] = temp;
-    }
-  }
-
-  // A matrix transpose turns each row into a column
-  private void transpose(int[][] matrix) {
-    for (int row = 0; row < matrix.length; row++) {
-      for (int col = row + 1; col < matrix[0].length; col++) {
-        int valueInUpperTriangle = matrix[row][col];
-        int valueInLowerTriangle = matrix[col][row];
-
-        matrix[row][col] = valueInLowerTriangle;
-        matrix[col][row] = valueInUpperTriangle;
       }
     }
   }
